working om accuracy

+++ Golf Game/src/Club.java
@@ -4,10 +4,10 @@ public class Club
 	{
 		private String name;
 		private int distance;
-		private double accuracy;
+		private int accuracy;
 		private boolean favorite;
 		
-		public Club(String n, int d, double a, boolean f)
+		public Club(String n, int d, int a, boolean f)
 		{
 			name=n;
 			distance=d;
@@ -35,12 +35,12 @@ public class Club
 				this.distance = distance;
 			}
 
-		public double getAccuracy()
+		public int getAccuracy()
 			{
 				return accuracy;
 			}
 
-		public void setAccuracy(double accuracy)
+		public void setAccuracy(int accuracy)
 			{
 				this.accuracy = accuracy;
 			}
@@ -59,19 +59,19 @@ public class Club
 		
 		public static void addClub()
 		{
-			club.add(new Club("Sand", 80, 80.0, false));
-			club.add(new Club("Approach", 110, 78.0, false));
-			club.add(new Club("Pitch", 135, 76.0, false));
-			club.add(new Club("9 Iron", 150, 75.0, false));
-			club.add(new Club("8 Iron", 160, 74.0, false));
-			club.add(new Club("7 Iron", 170, 73.0, false));
-			club.add(new Club("6 Iron", 180, 70.0, false));
-			club.add(new Club("5 Iron", 190, 68.0, false));
-			club.add(new Club("4 Iron", 200, 65.0, false));
-			club.add(new Club("Hybrid", 210, 70.0, false));
-			club.add(new Club("5 Wood", 220, 62.0, false));
-			club.add(new Club("3 Wood", 240, 60.0, false));
-			club.add(new Club("Driver", 260, 50.0, false));
+			club.add(new Club("Sand", 80, 80, false));
+			club.add(new Club("Approach", 110, 78, false));
+			club.add(new Club("Pitch", 135, 76, false));
+			club.add(new Club("9 Iron", 150, 75, false));
+			club.add(new Club("8 Iron", 160, 74, false));
+			club.add(new Club("7 Iron", 170, 73, false));
+			club.add(new Club("6 Iron", 180, 70, false));
+			club.add(new Club("5 Iron", 190, 68, false));
+			club.add(new Club("4 Iron", 200, 65, false));
+			club.add(new Club("Hybrid", 210, 70, false));
+			club.add(new Club("5 Wood", 220, 62, false));
+			club.add(new Club("3 Wood", 240, 60, false));
+			club.add(new Club("Driver", 260, 50, false));
 			
 		
 		}
